requirements_oauth.txt
# Обновленный список зависимостей с поддержкой OAuth
Flask==2.1.2
Flask-Caching==2.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==2.5.1
Werkzeug==2.3.7
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.7
python-dotenv==1.1.1
SQLAlchemy==1.4.46
blinker==1.9.0
colorama==0.4.6
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
requests==2.32.5
pillow==11.3.0
# OAuth Dependencies
Authlib==1.3.0
requests-oauthlib==2.0.0
cryptography>=3.4.8